Move profile-progress selectors to common
